+++ docs/changelog.md
@@ -49,6 +49,12 @@
 
 ## 更新说明
 
+1.8.47 add: add option_cffex_sz50_list_sina interface
+
+    1. 新增 option_cffex_sz50_list_sina 接口
+    2. 新增 option_cffex_sz50_spot_sina 接口
+    3. 新增 option_cffex_sz50_daily_sina 接口
+
 1.8.46 fix: fix get_roll_yield_bar interface
 
     1. 修复 get_roll_yield_bar 接口
@@ -1726,6 +1732,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.8.47 add: add option_cffex_sz50_list_sina interface
 
 1.8.46 fix: fix get_roll_yield_bar interface